Commit of the command handling class.
This is a rough draft of the command handling class. The help function will be changed as we progress, but I included just a sample statement for now. The class itself is named "cmd_handle." The function names are "exit_program" and "help." This commit was made at 11:20 A.M. on 11/21/2024.